unify outputs; add tests

+++ src/data/queries/api.py
@@ -8,7 +8,21 @@ from src.data.queries import curl as cq
 
 
 def get_by_name(name):
-    return API.get(name=name)
+    try:
+        return API.get(name=name)
+    except:
+        return None
+
+
+def get_by_id(id):
+    try:
+        return API.get(id=id)
+    except:
+        return None
+
+
+def delete(id):
+    return API.delete().where(API.id==id).execute()
 
 
 def list_apis():
@@ -17,6 +31,9 @@ def list_apis():
     
 
 def new_api(name, current=False):
+    api = get_by_name(name)
+    if api:
+        return None
     id = str(uuid4()).replace('-', '')
     date_current = datetime.now() if current else datetime(1970, 1, 1)
     api = API.create(id=id, name=name, date_current=date_current)
@@ -48,22 +65,36 @@ def get_curls(api):
 
 def add_to_api(api, curl_id):
     id = str(uuid4()).replace('-', '')
-    curl = Curl.get_by_id(curl_id)
-    APICurlJoin.create(id=id, api=api, curl=curl)
-    return True
+    qs = [cq.get_by_id, cq.get_by_name]
+    for q in qs:
+        curl = q(curl_id)
+        if curl:
+            APICurlJoin.create(id=id, api=api, curl=curl)
+            return True
+    return False
 
 
-def delete_from_api(api, curl_id):
-    curl = Curl.get_by_id(curl_id)
-    q = APICurlJoin.delete().where(APICurlJoin.api_id==api.id, APICurlJoin.curl_id==curl.id)
-    q.execute()
+def remove_from_api(api, curl_id):
+    qs = [cq.get_by_id, cq.get_by_name]
+    for q in qs:
+        curl = q(curl_id)
+        if curl:
+            q = APICurlJoin.delete().where(APICurlJoin.api_id==api.id, APICurlJoin.curl_id==curl.id)
+            q.execute()
+            return True
+    return False
 
 
 def to_json(api):
     curls = get_curls(api)
     result = {
+        "id": api.id,
         "name": api.name,
         "date_created": api.date_created.isoformat(),
         "curls": [cq.to_json(c) for c in curls]
     }
     return result
+
+
+def delete_all():
+    return API.delete().execute()
